# MAEBACKEND/Endpoints/openapi_specification.yaml
openapi: 3.0.3
info:
  title: EasyRent API - Sistema Completo de Marketplace Inmobiliario
  description: |
    API completa para el marketplace inmobiliario EasyRent con todas las funcionalidades:
    - 🔐 **Autenticación**: JWT + API Keys (Compatible con base de datos PostgreSQL)
    - 👥 **Gestión de usuarios y agencias**
    - 🏠 **Listings y multimedia**
    - 🔍 **Búsqueda avanzada y filtros**
    - 💎 **Sistema de planes y facturación**
    - 📊 **Analytics y reportes**
    - ✅ **Verificación y moderación**
    - 🔗 **Integraciones externas**
    
    ## Tecnologías
    - **Framework**: FastAPI + Python 3.11+
    - **Base de datos**: PostgreSQL 17
    - **Cache**: Redis
    - **Message Broker**: Apache Kafka
    - **Autenticación**: JWT (Bearer Token)
    
    ## Notas de Implementación
    - **Autenticación Firebase**: Sistema de autenticación completo con Firebase Authentication
    - **Flujo de autenticación**: Frontend → Firebase → Backend verification → JWT tokens
    - **Sin manejo de passwords**: Toda la gestión de credenciales delegada a Firebase
    - **Soporte híbrido**: Funciona con o sin Firebase configurado (modo desarrollo)
    - **Schemas actualizados**: Los modelos de datos coinciden con la implementación real en FastAPI
    
    ## Autenticación Firebase
    Para autenticarse, el cliente debe:
    1. Autenticarse con Firebase (frontend)
    2. Obtener el Firebase ID token
    3. Enviar el token al endpoint `/auth/login`
...
    
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Listar usuarios (admin)
      description: Obtiene una lista paginada de usuarios con filtros (solo administradores)
      operationId: list_users
      parameters:
        - name: role
          in: query
          description: Filtrar por rol
          schema:
            type: string
            enum: [user, tenant, landlord, agent, admin]
        - name: search
          in: query
          description: Buscar por nombre o email
          schema:
            type: string
            maxLength: 100
        - name: page
          in: query
          description: Número de página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Elementos por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
        '401':
          description: No autorizado
        '403':
          description: Permisos insuficientes (solo administradores)
        '500':
          description: Error interno del servidor

    post:
      tags: [Users]
      summary: Crear usuario (admin)
      description: Crea un nuevo usuario (solo administradores)
      operationId: create_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
        '403':
          description: Permisos insuficientes
        '409':
          description: Email ya existe
        '500':
          description: Error interno del servidor
...
  /users/{user_id}:
    get:
      tags: [Users]
      summary: Obtener usuario por ID
      description: Obtiene la información pública de un usuario (o completa si es admin)
      operationId: get_user_by_id
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Datos del usuario
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/UserDetailResponse'
                  - $ref: '#/components/schemas/UserPublicResponse'
                description: UserDetailResponse si es admin, UserPublicResponse si es usuario normal
        '404':
          description: Usuario no encontrado
        '500':
          description: Error interno del servidor

    put:
      tags: [Users]
      summary: Actualizar usuario (admin)
      description: Actualiza un usuario específico (solo administradores)
      operationId: update_user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
        '403':
          description: Permisos insuficientes
        '404':
          description: Usuario no encontrado
        '500':
          description: Error interno del servidor

    delete:
      tags: [Users]
      summary: Eliminar usuario (admin)
      description: Elimina un usuario específico (solo administradores)
      operationId: delete_user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuario eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: No se puede eliminar a sí mismo
        '401':
          description: No autorizado
        '403':
          description: Permisos insuficientes
        '404':
          description: Usuario no encontrado
        '500':
          description: Error interno del servidor
...
  /agencies:
    get:
      tags: [Agencies]
      summary: Listar agencias
      description: Obtiene una lista de agencias inmobiliarias con filtros opcionales
      operationId: list_agencies
      security: []
      parameters:
        - name: city
          in: query
          description: Filtrar por ciudad
          schema:
            type: string
            example: "Lima"
        - name: verified
          in: query
          description: Solo agencias verificadas
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: Lista de agencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgencyResponse'
        '500':
          description: Error interno del servidor

    post:
      tags: [Agencies]
      summary: Crear agencia
      description: Crea una nueva agencia inmobiliaria
      operationId: create_agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgencyRequest'
      responses:
        '201':
          description: Agencia creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyResponse'
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor
  /agencies/{agency_id}:
    get:
      tags: [Agencies]
      summary: Obtener agencia por ID
      description: Obtiene los detalles de una agencia específica
      operationId: get_agency
      security: []
      parameters:
        - name: agency_id
          in: path
          required: true
          description: ID único de la agencia
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Datos de la agencia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyResponse'
        '404':
          description: Agencia no encontrada
        '500':
          description: Error interno del servidor

    put:
      tags: [Agencies]
      summary: Actualizar agencia
      description: Actualiza los datos de una agencia existente
      operationId: update_agency
      parameters:
        - name: agency_id
          in: path
          required: true
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgencyRequest'
      responses:
        '200':
          description: Agencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyResponse'
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado
        '404':
          description: Agencia no encontrada
        '500':
          description: Error interno del servidor

    delete:
      tags: [Agencies]
      summary: Eliminar agencia
      description: Elimina una agencia del sistema
      operationId: delete_agency
      parameters:
        - name: agency_id
          in: path
          required: true
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Agencia eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: No autorizado
        '404':
          description: Agencia no encontrada
        '500':
          description: Error interno del servidor
...
  /listings:
    get:
      tags: [Listings]
      summary: Listar propiedades
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: operation_type
          in: query
          schema:
            type: string
            enum: [sale, rent, temp_rent, pre_sale]
        - name: property_type
          in: query
          schema:
            type: string
            enum: [house, apartment, studio, room, office, commercial, land, warehouse, garage, other]
        - name: city
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: min_price
          in: query
          schema:
            type: number
        - name: max_price
          in: query
          schema:
            type: number
        - name: bedrooms
          in: query
          schema:
            type: integer
        - name: bathrooms
          in: query
          schema:
            type: integer
        - name: min_area
          in: query
          schema:
            type: number
        - name: max_area
          in: query
          schema:
            type: number
        - name: min_age_years
          in: query
          schema:
            type: integer
            minimum: 0
            description: Antigüedad mínima de la vivienda en años
        - name: max_age_years
          in: query
          schema:
            type: integer
            minimum: 0
            description: Antigüedad máxima de la vivienda en años
        - name: features
          in: query
          schema:
            type: array
            items:
              type: string
        - name: amenities
          in: query
          schema:
            type: array
            items:
              type: string
        - name: verified
          in: query
          schema:
            type: boolean
        - name: pet_friendly
          in: query
          schema:
            type: boolean
            description: Filtrar por propiedades que permiten mascotas
        - name: furnished
          in: query
          schema:
            type: boolean
            description: Filtrar por propiedades amuebladas/amobladas
        - name: rental_mode
          in: query
          schema:
            type: string
            enum: [full_property, private_room, shared_room]
            description: Modalidad de alquiler
        - name: airbnb_eligible
          in: query
          schema:
            type: boolean
            description: Solo propiedades elegibles para Airbnb
        - name: min_airbnb_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
            description: Score mínimo de elegibilidad Airbnb (0-100)
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, area_asc, area_desc, date_asc, date_desc, age_asc, age_desc, relevance]
      responses:
        '200':
          description: Lista de propiedades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedListings'

    post:
      tags: [Listings]
      summary: Crear nueva propiedad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Propiedad creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
  /listings/{listing_id}:
    get:
      tags: [Listings]
      summary: Obtener propiedad por ID
      security: []
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Datos de la propiedad
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'

    put:
      tags: [Listings]
      summary: Actualizar propiedad
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListingRequest'
      responses:
        '200':
          description: Propiedad actualizada

    delete:
      tags: [Listings]
      summary: Eliminar propiedad
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Propiedad eliminada
...
  /search:
    get:
      tags: [Search]
      summary: Búsqueda general
      security: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
            description: Texto de búsqueda
        - name: location
          in: query
          schema:
            type: string
            description: Ubicación (ciudad, distrito, etc.)
        - name: lat
          in: query
          schema:
            type: number
            description: Latitud para búsqueda por proximidad
        - name: lng
          in: query
          schema:
            type: number
            description: Longitud para búsqueda por proximidad
        - name: radius
          in: query
          schema:
            type: number
            description: Radio en kilómetros
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
...
  /api-keys:
    get:
      tags: [Developer Tools]
      summary: Listar API keys
      responses:
        '200':
          description: Lista de API keys

    post:
      tags: [Developer Tools]
      summary: Crear API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [read, write, admin]
                rate_limit:
                  type: integer
                  default: 1000
                expires_at:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        '201':
          description: API key creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  key:
                    type: string
                  name:
                    type: string
                  scopes:
                    type: array
                    items:
                      type: string
...
  /api-keys/{key_id}/regenerate:
    post:
      tags: [Developer Tools]
      summary: Regenerar API key
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API key regenerada
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string

  # ===========================================
  # SYSTEM UTILITIES
  # ===========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
        
        **Ejemplo**: `Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
...